Fixxor for first build

+++ kernel/sched/fair.c
@@ -4129,11 +4129,11 @@ unsigned long __weak arch_max_cpu_capacity(int cpu)
 	return SCHED_POWER_SCALE;
 }
 
-unsigned long __weak arch_get_cpu_capacity(int cpu)
+int __weak arch_get_cpu_capacity(int cpu)
 {
 	return SCHED_POWER_SCALE;
 }
-unsigned long __weak arch_get_max_cpu_capacity(int cpu)
+int __weak arch_get_max_cpu_capacity(int cpu)
 {
 	return SCHED_POWER_SCALE;
 }
